gomtree: return exit status != 0 on error
This was broken during the refactor for "gomtree validate" in commit
83c9fdb78bf7 ("refactor: prefactor for adding new subcommands"),
resulting in any code that relied on our exit code to silently treat all
errors as non-fatal.

Our tests did not catch this due to a quirky POSIX-ism with regards to
"! cmd" and "set -e" which is fixed in a follow-up patch.

Fixes: 83c9fdb78bf7 ("refactor: prefactor for adding new subcommands")
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>